+++ src/commandGroups/all.ts
@@ -4,6 +4,7 @@ import { commandTable as otherShells } from "./othershells.js";
 import { commandTable as processCmds } from "./processCommands.js";
 import { commandTable as streamsC } from "./streams.js";
 import { commandTable as thatShell } from "./thatShell.js";
+import { commandTable as webGroup } from "./webGroup.js";
 import { commandTable as uptime} from "../commands/uptime.js";
 import { commandTable as meow} from "../commands/meow.js";
 import { commandCollection } from "../tools/commandCollection.js";
@@ -14,10 +15,9 @@ const commandTable = {
     ...processCmds,
     ...streamsC,
     ...thatShell,
+    ...webGroup,
     ...uptime,
-    ...meow,
-    
-
+    ...meow
 };
 
 const compounds = cmdTableToCommandCompounts(commandTable)