+++ src/commandGroups/webGroup.ts
@@ -1,0 +1,13 @@
+import { commandTable as gods} from "../commands/gods.js";
+import { commandCollection } from "../tools/commandCollection.js";
+import { cmdTableToCommandCompounts } from "../tools/commandCreatorTools.js";
+
+const commandTable = {
+    ...gods,
+};
+
+const compounds = cmdTableToCommandCompounts(commandTable);
+
+const collection = new commandCollection(commandTable);
+
+export {commandTable, compounds, collection}