+++ src/in.ts
@@ -39,13 +39,48 @@ const allowedKeysToWrite: string = "abcdefghijklmnopqrstuxwvyz" + "abcdefghijklm
 //     // commandInternalExec(text);
 // }
 
-function inHandler(data: any, ses: terminalSession){
+async function inHandler(data: any, ses: terminalSession){
     //    console.log(data);
     if(data){
         // control + X
         if(data.includes("\u0018")){
-            ses.text += "§";
+            if(ses.relativeTextboxPos !== 0){
+                // create a new text with that key
+                ses.text = ses.text.slice(0, ses.relativeTextboxPos) 
+                + '§' + ses.text.slice(ses.relativeTextboxPos);
+
+                
+                // hideCursor();
+                // process.stdout.write("\r");
+
+                // ses.out.write(hideCursorCODE);
+                // ses.out.write(
+                //     clearEntireLineCODE + "\r" +
+                //     formatPrintTextbox(ses.text, ses) +
+                //     cursorRel(ses.relativeTextboxPos, 0) +
+                //      showCursorCODE
+                // );
+                printTextBox(ses);
+
+                // ses.out.write(hideCursorCODE + clearEntireLineCODE + "\r");
+
+                // // printViewTextbox(ses.text, ses.out);
+
+                // ses.out.write(cursorRel(ses.relativeTextboxPos, 0) + showCursorCODE);
+
+                // process.stdout.moveCursor(relativePos, 0);
+                // showCursor();
+
+                return;
+            }
+
+            ses.text += '§';
+            // ses.out.write(data); 
             printTextBox(ses);
+
+
+            // ses.text += "§";
+            // printTextBox(ses);
             return;
         }
 
@@ -212,7 +247,7 @@ function inHandler(data: any, ses: terminalSession){
 
             // command exec
             // commandInternalExec(text, ses);
-                commandInternalExec(text, {
+                await commandInternalExec(text, {
                     logNode: "console",
                     silent: false,
                     onlyReturn: false,
@@ -388,8 +423,8 @@ function setupInHandlerListener(ses: terminalSession, setupStream: boolean = tru
 
     ses.in.addListener(
         "data",
-        (chunk) => {
-            inHandler(chunk, ses);
+        async (chunk) => {
+            await inHandler(chunk, ses);
         }
     );
 }